+++ openquake/server/tests/helpers.py
@@ -98,45 +98,6 @@ def run_job(cfg, exports='xml,csv', hazard_calculation_id=None, **params):
     return engine.run_calc(job_id, oqparam, 'error', logfile, exports)
 
 
-class ConfigTestCase(object):
-    """Class which contains various configuration- and environment-related
-    testing helpers."""
-
-    def setup_config(self):
-        self.orig_env = os.environ.copy()
-        os.environ.clear()
-        # Move the local configuration file out of the way if it exists.
-        # Otherwise the tests that follow will break.
-        local_path = config.cfg.PKG_PATH
-        if os.path.isfile(local_path):
-            shutil.move(local_path, "%s.test_bakk" % local_path)
-
-    def teardown_config(self):
-        os.environ.clear()
-        os.environ.update(self.orig_env)
-        # Move the local configuration file back into place if it was stashed
-        # away.
-        local_path = config.cfg.PKG_PATH
-        if os.path.isfile("%s.test_bakk" % local_path):
-            shutil.move("%s.test_bakk" % local_path, local_path)
-        config.cfg.cfg.clear()
-        config.cfg._load_from_file()
-
-    def prepare_config(self, section, data=None):
-        """Set up a configuration with the given `max_mem` value."""
-        if data is not None:
-            data = '\n'.join("%s=%s" % item for item in data.items())
-            content = """
-                [%s]
-                %s""" % (section, data)
-        else:
-            content = ""
-        site_path = touch(content=textwrap.dedent(content))
-        os.environ["OQ_CONFIG_FILE"] = site_path
-        config.cfg.cfg.clear()
-        config.cfg._load_from_file()
-
-
 def random_string(length=16):
     """Generate a random string of the given length."""
     result = ""
